Update and rename plugin.php to SingleID.php

+++ SingleID.php
@@ -3,7 +3,7 @@ header("Access-Control-Allow-Origin: *");
 
 
 /*
- * SINGLEID WEB PLUGIN
+ * SingleID WEB PLUGIN -> https://github.com/SingleID/web-plugin/
  * Date: 2015-02 from SingleID Inc.
  * 
  * To use the plugin on your site please upload this file to your web root directory.
@@ -11,8 +11,7 @@ header("Access-Control-Allow-Origin: *");
  * <script src="https://ajax.googleapis.com/ajax/libs/jquery/1.6.1/jquery.min.js"></script>
  *
  * The next step is to insert the SingleId Button. Place this line of code to the place of your site where you want to place the button:
- * <iframe src="http://www.example.com/plugin.php?op=init" width="200" height="80" frameborder="0"></iframe>
- * where www.example.com is your domain
+ * <iframe src="SingleID.php?op=init" width="270" height="80" frameborder="0"></iframe>
  *
  * On line 26 and 27 you must use your site logo and site name.
  * On line 28 you can change the data that you want to receive
@@ -30,32 +29,19 @@ header("Access-Control-Allow-Origin: *");
  
  
  
-define("LOGO_URL", "http://www.singleid.com/img/logonew.png");
-define("SITE_NAME", "Basic test");
-define("requested_data", "1,2,3,4");
-define("billing_key", ""); 	// You have to request this key from www.singleid.com if requested_data is different from "1"
-define("admin_contact", "xxxxxx@singleid.com");
-
-
-function is_SingleID($val){
-
-if (strlen($val) == '8'){  
-	return (bool)preg_match("/[0-9a-f]{8}/i", $val);
-}else if (strlen($val) == '7'){  	// crockford mode base 32
-	return (bool)preg_match("/[0-9a-z]{7}/i", $val);
-}else{
-	return false;
-}
-
-}
+define("LOGO_URL", "http://avatars0.githubusercontent.com/u/10206030?v=3&s=40");
+define("SITE_NAME", "GITHUB");
+define("requested_data", "1");
+define("billing_key", ""); 						// You have to request this key from www.singleid.com if requested_data is different from "1"
+define("admin_contact", "");					// You have to set this field only if requested_data is different from "1"
 
 
 
-// first check !
+// first of all
 
-if(!is_writable('userdata/')){
-	print '<p>no write permission!</p>';
+if (!is_writable('userdata/')) {
 	error_log('no permission for userdata/ folder TRY -> sudo chmod 0777 userdata/ -R ');
+	die('<p>no write permission!</p>');
 }
 
 
@@ -64,26 +50,16 @@ if(!is_writable('userdata/')){
 
 	session_start();
 
+	if (isset($_POST['UTID'])) { // to change !
+		$op = 'response'; // When some device is sending data !
+	} else {
+		$op = $_REQUEST['op'];
+	}
 
 
-
-		if(isset($_POST['UTID'])) {
-			$op = 'response'; // When some device is sending data !
-		} else {
-			$op = $_REQUEST['op'];
-		}
-
-
-
-
-
-if($op == 'init'){ // Where all begin ( from browser user )
-
-		$_SESSION['SingleID']['hash'] = md5( microtime().md5($_SERVER['HTTP_USER_AGENT'].mt_rand(1, mt_getrandmax())).$_SERVER['REMOTE_ADDR'].$_SERVER['SCRIPT_FILENAME'].mt_rand(1, mt_getrandmax()) );
+if ($op == 'init') { // Where all begin ( from browser user )
+		error_log('button is displayed');
 		
-		
-		$_SESSION['SingleID'][$_SESSION['SingleID']['hash']]['has_response'] = 0;
-		$_SESSION['SingleID'][$_SESSION['SingleID']['hash']]['is_sended'] = 0;
 
 		?>
 		<!DOCTYPE html>
@@ -91,21 +67,21 @@ if($op == 'init'){ // Where all begin ( from browser user )
 			<head>
 				<title>SingleID iframed button</title>
 				<meta charset="utf-8">
-				<link rel="stylesheet" href="https://app.singleid.com/button/plugin/css/main_sheet.css">
+				<link rel="stylesheet" href="css/SingleID/SingleID.css">
 				<script src="https://ajax.googleapis.com/ajax/libs/jquery/1.6.1/jquery.min.js"></script>				
 				<script src="js/plugin.js"></script>
 			</head>
 			<body>
 		   		<div class="singleid_button_wrap singleid_pointer">
-	        	    <div class="single_text_single_id">SingleID</div>
-	                <div class="icon_box_single_id"><img src="https://app.singleid.com/button/plugin/img/key2.jpg" alt="" /></div>
+	        	    <div class="single_text_single_id">login with your SingleID</div>
+	                <div class="icon_box_single_id"><img src="css/SingleID/SingleID_logo_key.jpg" alt="No more form filling with SingleID" /></div>
 
 	            	<div class="white_back_single_id singleid_invisible">
 	                	<input class="singleid_styled_input" name="SingleID" type="text" value="" maxlength="8" />
 	                	<button type="button" class="icon_box_go" onClick="sid_sendData();">go</button>
 	            	</div>
 	                <div class="singleid_waiting singleid_invisible">waiting for data</div>
-				    <a href="https://www.singleid.com" title="SingleID is available for Android, iPhone and Windows Phone"><div class="free_text_single_id">Get SingleID now!</div>
+				    <a href="https://www.singleid.com" target="_top" title="SingleID is available for Android, iPhone and Windows Phone"><div class="free_text_single_id">Get SingleID now!</div>
 					</a>
 			    </div>
 			   </body>
@@ -142,6 +118,21 @@ if($op == 'init'){ // Where all begin ( from browser user )
 	
 }elseif($op == 'send'){ 	// From browser 
 							// here start the request to forward 
+							
+	$securitydata = '<html><h1>Silence is gold</h1></html>';	 // just to be extra sure that nobody could browse this folder that form some minutes is full of sensitive data
+	$fp = fopen('userdata/index.html', 'w');
+	fwrite($fp, $securitydata);
+	fclose($fp);
+							
+							
+	error_log('user has clicked go');
+	
+	$_SESSION['SingleID']['hash'] = md5( microtime().md5($_SERVER['HTTP_USER_AGENT'].mt_rand(1, mt_getrandmax())).$_SERVER['REMOTE_ADDR'].$_SERVER['SCRIPT_FILENAME'].mt_rand(1, mt_getrandmax()) );
+	$_SESSION['SingleID'][$_SESSION['SingleID']['hash']]['has_response'] = 0;
+	$_SESSION['SingleID'][$_SESSION['SingleID']['hash']]['is_sended'] = 0;
+		
+		
+		
 
 	if (is_SingleID($_POST['single_id'])) {
 		
@@ -159,6 +150,7 @@ if($op == 'init'){ // Where all begin ( from browser user )
 					error_log('send 2 '.$ssl); // in questo caso devo bloccare il pulsante !!!! [TODO]
 				}
 			
+			//error_log('AFTERIF requested_data='.requested_data.'||| ssl'.$ssl);	
 			//set POST variables
 			$url = 'https://app.singleid.com/';
 
@@ -195,16 +187,40 @@ if($op == 'init'){ // Where all begin ( from browser user )
 			$responseInfo = curl_getinfo($ch);
 
 			$ServerReply = json_decode($result, true);
+
+
+				// for debug purposes only
+						/*
+						if (empty($result)) {
+							// some kind of an error happened
+							die(curl_error($ch));
+							curl_close($ch); // close cURL handler
+						} else {
+							$info = curl_getinfo($ch);
+							curl_close($ch); // close cURL handler
+
+							if (empty($info['http_code'])) {
+									die("No HTTP code was returned");
+							} else {
+								// load the HTTP codes
+								$http_codes = parse_ini_file("path/to/the/ini/file/I/pasted/above");
+							   
+								// echo results
+								echo "The server responded: <br />";
+								echo $info['http_code'] . " " . $http_codes[$info['http_code']];
+							}
+
+						}
+						*/
+
+
 			
-
-
-			//close connection
-			curl_close($ch);
+			curl_close($ch); //close connection because we are brave 
 
 			$_SESSION['SingleID'][$_SESSION['SingleID']['hash']]['is_sended'] = time();
 		}
 		
-			if ($ServerReply['Reply'] <> 'ok'){
+			if ($ServerReply['Reply'] <> 'ok') {
 			error_log($ServerReply['PopupTitle'].' '. $ServerReply['Popup']);
 			die($ServerReply['PopupTitle']);
 			}
@@ -217,13 +233,16 @@ if($op == 'init'){ // Where all begin ( from browser user )
 	
 } elseif($op == 'response') {
 	// This happen when a Device has sent something ( $_POST[UTID]) 
-	
+	if (!is_md5($_POST['UTID'])){
+		die('Wrong data received!');
+		unset($_SESSION['SingleID']); // leave the system clean for better security
+	}
 	
 		// TODO we need also to remove from this folder all the files older than 5 minutes for security reason !!!
 		$path = 'userdata/';
 		  if ($handle = opendir($path)) {
 			 while (false !== ($file = readdir($handle))) {
-				if ((time()-filectime($path.$file)) >= 300) {  
+				if ((time()-filectime($path.$file)) >= 300) {
 				   if (preg_match('/\.SingleID.txt$/i', $file)) {
 					  unlink($path.$file);
 				   }
@@ -233,6 +252,7 @@ if($op == 'init'){ // Where all begin ( from browser user )
 	   
 	   
 	   
+	error_log('DEVICE_RESP'.serialize($_POST));
 
 	if ($_POST['Ecom_payment_mode'] != 'paypal')	{
 		$_POST['Ecom_payment_mode'] = $_POST['Ecom_payment_card_type'];
@@ -251,14 +271,20 @@ if($op == 'init'){ // Where all begin ( from browser user )
 	
 	
 } elseif($op == 'getdata') {
-	error_log('getdata========');
+	
+	if (!is_md5($_SESSION['SingleID']['hash'])){
+		die('Wrong data to recover');
+		unset($_SESSION['SingleID']); // leave the system clean for better security
+	}
+	
+	error_log('getdata========'.$_SESSION['SingleID']['hash']);
 	$filetarget = 'userdata/'.$_SESSION['SingleID']['hash'].'.SingleID.txt';
 	$data = unserialize(gzuncompress(file_get_contents($filetarget)));
 	$data['ALREADY_REGISTERED'] = 1; // to understand security of this...
-
-				// TODO TOCHECK 
-				// THIS IS THE MOST DELICATE POINT !
-
+	
+									 // if this value is set to 1
+									 // The javascript know that is not a form filling example but the data has been processed from PHP
+	// start from here we need
 	
 	// but if I am already a registered user ?
 	
@@ -267,22 +293,63 @@ if($op == 'init'){ // Where all begin ( from browser user )
 	
 	
 	 // CODE FOR WEBSITE OWNER  START =======
-	
-	 
 
-			// the SingleID is already present in the profile ?
+	
+	/* data received with a personal profile 
+			"Pers_title",
+			"Pers_first_name",
+			"Pers_middle_name",
+			"Pers_last_name",
+			"Pers_birthdate",
+			"Pers_gender",
+			"Pers_postal_street_line_1",
+			"Pers_postal_street_line_2",
+			"Pers_postal_street_line_3",
+			"Pers_postal_city",
+			"Pers_postal_postalcode",
+			"Pers_postal_stateprov",
+			"Pers_postal_countrycode",
+			"Pers_telecom_fixed_phone",
+			"Pers_telecom_mobile_phone",
+			"Pers_first_email",
+			"Pers_skype",
+			"Pers_first_language",
+			"Pers_second_language",
+			"Pers_contact_preferred_mode",
+			"Pers_newsletter_agree"
+      */
+      
+      
+                            
+	// Possibilities
+	
+	// User not present in DB with SingleID and not present with the same Email !
+		
+		// we try to auto register if the profile contains a minimum of data
+		
+	
+	// User not present in DB with SingleID BUT with email already present
+	
+		// The system should send a confirmation link to the email already present to enable the association from the user !
+		
+		
+	// User present in DB with SingleID but account disabled
+		
+		// display error
+	
+	// User present in DB with SingleID 
+	
+		// which profile we accept ? (personal/companies) ?
+		
+		// We need to launch an update query of the user data ( if met the minimum profile data request )
 			
-				// if yes we need to merge this profile but how ???
-				
-				
-			// if the email is not present we need to register the user
-	
-				// IN DEVELOPMENT
+			// warning we had not to replace the data that we do not the user could replace ! ( name / surname )
+			
+			// Warning... if the user now has just logged with a personal profile and then with a company profile ?
 	
 	
-	
-		// so i need to update the allowed value and then mark it as logged in
-	
+
+
 	// CODE FOR WEBSITE OWNER END =====
 	
 	
@@ -297,12 +364,12 @@ if($op == 'init'){ // Where all begin ( from browser user )
 	// See if it exists before attempting deletion on it
 	if (file_exists($filetarget)) {
     unlink($filetarget); // Delete now
-    unset($_SESSION['SingleID']); // we love to clean the system after use 
+    unset($_SESSION['SingleID']); // leave the system clean for better security
 
 	} 
 	// See if it exists again to be sure it was removed
 	if (file_exists($filetarget)) {
-		error_log('Problem deleting sensitive SingleID files. Check your permission ! ' . $filetarget);
+		error_log('Problem deleting. Check your permission ! ' . $filetarget);
 	}
 	die;
 	
@@ -312,19 +379,23 @@ if($op == 'init'){ // Where all begin ( from browser user )
 	
 	// request made from browser !
 	
+	error_log('refresh');
+
+	
+	//if($_SESSION['SingleID'][$_SESSION['SingleID']['hash']]['is_sended'] > 0)
+	//{
 		$file = 'userdata/'.$_SESSION['SingleID']['hash'].'.SingleID.txt';
 		//$file300 = 'userdata/'.$_SESSION['SingleID']['hash'].'.300.SingleID.txt';
 		//$dif = time() - $_SESSION['SingleID'][$_SESSION['SingleID']['hash']]['is_sended'];
 
 		if($dif > 60){
-			error_log('400-0000'); // TODO 
+			error_log('400-0000');
 			print 400; // too much time is passed
 			
 		}else if( is_file($file) ){ // the file exist !
-			 
-				error_log('200-0000');
+
+				error_log('DEBUG: DATA RECEIVED FROM DEVICE');
 				print '200'; // the post data has been received from the device so we launch the JS to populate the fields
-			//}
 		}
 		else
 		{
@@ -335,8 +406,27 @@ if($op == 'init'){ // Where all begin ( from browser user )
 }
 
 
+// debug info
+echo '<p>This script must be embedded into a page to work correctly</p>';
 
 
+
+
+
+
+function is_SingleID($val){
+	if (strlen($val) == '8'){  
+		return (bool)preg_match("/[0-9a-f]{8}/i", $val);
+		}else if (strlen($val) == '7'){  	// crockford mode base 32
+		return (bool)preg_match("/[0-9a-z]{7}/i", $val);
+		}else{
+		return false;
+	}
+}
+
+function is_md5($val) {
+	return (bool)preg_match("/[0-9a-f]{32}/i", $val);
+}
   
   
 ?>
